MODINVSTOR-742 Move endpoint to another interface
