+++ src/main/java/org/folio/rest/impl/PrecedingSucceedingTitleAPI.java
@@ -1,18 +1,29 @@
 package org.folio.rest.impl;
 
+import static org.folio.rest.persist.PgUtil.postgresClient;
+import static org.folio.rest.tools.utils.ValidationHelper.createValidationErrorMessage;
+
+import java.util.List;
+import java.util.Map;
+
+import javax.ws.rs.core.Response;
+
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Context;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
+import io.vertx.core.Promise;
+import io.vertx.sqlclient.Row;
+import io.vertx.sqlclient.RowSet;
 import org.apache.commons.lang3.ObjectUtils;
+
 import org.folio.rest.annotations.Validate;
 import org.folio.rest.jaxrs.model.PrecedingSucceedingTitle;
 import org.folio.rest.jaxrs.model.PrecedingSucceedingTitles;
 import org.folio.rest.persist.PgUtil;
+import org.folio.rest.persist.PostgresClient;
+import org.folio.rest.tools.utils.MetadataUtil;
 import org.folio.rest.tools.utils.ValidationHelper;
-
-import javax.ws.rs.core.Response;
-import java.util.Map;
 
 public class PrecedingSucceedingTitleAPI implements org.folio.rest.jaxrs.resource.PrecedingSucceedingTitles {
 
@@ -37,6 +48,21 @@ public class PrecedingSucceedingTitleAPI implements org.folio.rest.jaxrs.resourc
     }
     PgUtil.post(PRECEDING_SUCCEEDING_TITLE_TABLE, entity, okapiHeaders, vertxContext,
       PostPrecedingSucceedingTitlesResponse.class, asyncResultHandler);
+  }
+
+  @Override
+  public void putPrecedingSucceedingTitles(String instanceId, PrecedingSucceedingTitles entity,
+                                           Map<String, String> okapiHeaders,
+                                           Handler<AsyncResult<Response>> asyncResultHandler, Context vertxContext) {
+    var titles = entity.getPrecedingSucceedingTitles();
+    boolean areValidTitles = validatePrecedingSucceedingTitles(titles, instanceId, asyncResultHandler);
+    if (areValidTitles) {
+      var cqlQuery = String.format("succeedingInstanceId==(%s) or precedingInstanceId==(%s)", instanceId, instanceId);
+      PgUtil.delete(PRECEDING_SUCCEEDING_TITLE_TABLE, cqlQuery, okapiHeaders, vertxContext,
+          PutPrecedingSucceedingTitlesResponse.class)
+        .compose(response -> saveCollection(titles, okapiHeaders, vertxContext))
+        .onComplete(asyncResultHandler);
+    }
   }
 
   @Validate
@@ -66,6 +92,56 @@ public class PrecedingSucceedingTitleAPI implements org.folio.rest.jaxrs.resourc
     }
     PgUtil.put(PRECEDING_SUCCEEDING_TITLE_TABLE, entity, precedingSucceedingTitleId, okapiHeaders, vertxContext,
       PutPrecedingSucceedingTitlesByPrecedingSucceedingTitleIdResponse.class, asyncResultHandler);
+  }
+
+  private boolean validatePrecedingSucceedingTitles(List<PrecedingSucceedingTitle> precedingSucceedingTitles,
+                                                    String instanceId,
+                                                    Handler<AsyncResult<Response>> asyncResultHandler) {
+    boolean areValidTitles = true;
+    for (PrecedingSucceedingTitle precedingSucceedingTitle : precedingSucceedingTitles) {
+      if (titleIsLinkedToInstanceId(precedingSucceedingTitle, instanceId)) {
+        var validationErrorMessage =
+          createValidationErrorMessage("precedingInstanceId or succeedingInstanceId", "",
+            String.format("The precedingInstanceId or succeedingInstanceId should contain instanceId [%s]", instanceId));
+        asyncResultHandler.handle(Future.succeededFuture(PutPrecedingSucceedingTitlesResponse
+          .respond422WithApplicationJson(validationErrorMessage)));
+        areValidTitles = false;
+      }
+    }
+    return areValidTitles;
+  }
+
+  private boolean titleIsLinkedToInstanceId(PrecedingSucceedingTitle precedingSucceedingTitle, String instanceId) {
+    return !instanceId.equals(precedingSucceedingTitle.getPrecedingInstanceId())
+      && !instanceId.equals(precedingSucceedingTitle.getSucceedingInstanceId());
+  }
+
+  private Future<Response> saveCollection(List<PrecedingSucceedingTitle> entities,
+                                          Map<String, String> okapiHeaders, Context vertxContext) {
+    Promise<Response> promise = Promise.promise();
+
+    try {
+      MetadataUtil.populateMetadata(entities, okapiHeaders);
+      PostgresClient postgresClient = postgresClient(vertxContext, okapiHeaders);
+      Handler<AsyncResult<RowSet<Row>>> replyHandler = result -> {
+        if (result.failed()) {
+          var errorMessage = result.cause().getMessage();
+          PutPrecedingSucceedingTitlesResponse response;
+          if (ValidationHelper.isFKViolation(errorMessage)) {
+            response = PutPrecedingSucceedingTitlesResponse.respond404WithTextPlain("Instance not found");
+          } else {
+            response = PutPrecedingSucceedingTitlesResponse.respond500WithTextPlain(errorMessage);
+          }
+          promise.complete(response);
+        } else {
+          promise.complete(PutPrecedingSucceedingTitlesResponse.respond204());
+        }
+      };
+      postgresClient.saveBatch(PRECEDING_SUCCEEDING_TITLE_TABLE, entities, replyHandler);
+    } catch (ReflectiveOperationException e) {
+      promise.complete(PutPrecedingSucceedingTitlesResponse.respond500WithTextPlain(e.getMessage()));
+    }
+    return promise.future();
   }
 
   private boolean precedingAndSucceedingInstanceEmpty(PrecedingSucceedingTitle entity) {
